new functionality: sharing with key, UI changes.
Sharing with the key:

The idea is that all drawings get a share key that can be shared with whatever
means (e.g. Google+ private posting). Then when the person who got the key
enters it, he/she gets access to that drawing - the email address is added to
the "shared" list. This way the people who have seen the drawing can be
tracked.

The inspiration was that you can share the key in e.g. Google+ private post.

UI changes: hopefully it is more usable now.

I added a separate "Save Drawing" button, which allows saving and deleting
drawings. Also from there you can create a new drawing ("save as").